Keep top bars fixated
The filter and group bar now stay where they are. However the bar on top of the group view sadly still moves.
But else the whole group view is now scrollable, and the topbars are usable from any scroll point.
Fixed some insets stuff in the custom flow-layout.
Now there is actually a need for Y-Padding and there are still edge cases where it breaks.